Fix width/height for trimmed images
The width and height reported by `identify` for trimmed PNG images is incorrect because the transparent border is included in the new page size and offset. For example, for a 32x32 trimmed image, `identify` reports a geometry of 34x34+1+1. Also, because PNG images support canvas metadata but other formats may not (JPEG), the trimmed file is always explicitly converted to PNG so that the new calculation always works correctly.